@@ -1,5 +1,5 @@
 /*
- * (C) Copyright [2021-2023] Hewlett Packard Enterprise Development LP
+ * (C) Copyright [2021-2024] Hewlett Packard Enterprise Development LP
  *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
@@ -25,7 +25,7 @@ package api
 import (
 	"encoding/json"
 	"errors"
-	"io/ioutil"
+	"io"
 	"net/http"
 
 	// base "github.com/Cray-HPE/hms-base"
@@ -48,7 +48,11 @@ func CreateTransition(w http.ResponseWriter, req *http.Request) {
 	var pb model.Passback
 	var parameters model.TransitionParameter
 	if req.Body != nil {
-		body, err := ioutil.ReadAll(req.Body)
+		body, err := io.ReadAll(req.Body)
+
+		// Not necessarily needed, but close request body anyways
+		req.Body.Close()
+
 		logger.Log.WithFields(logrus.Fields{"body": string(body)}).Trace("Printing request body")
 
 		if err != nil {
@@ -105,6 +109,13 @@ func GetTransitions(w http.ResponseWriter, req *http.Request) {
 	if _, ok := params["transitionID"]; ok {
 		//parse uuid and if its good then call GetTransition
 		pb = GetUUIDFromVars("transitionID", req)
+
+		// Drain and close request body to ensure connection reuse
+		if (req.Body != nil) {
+			_, _ = io.Copy(io.Discard, req.Body)	// drain in case partially read
+			req.Body.Close()
+		}
+
 		if pb.IsError {
 			WriteHeaders(w, pb)
 			return
@@ -113,6 +124,12 @@ func GetTransitions(w http.ResponseWriter, req *http.Request) {
 		pb = domain.GetTransition(transitionID)
 
 	} else {
+		// Drain and close request body to ensure connection reuse
+		if (req.Body != nil) {
+			_, _ = io.Copy(io.Discard, req.Body)
+			req.Body.Close()
+		}
+
 		pb = domain.GetTransitionStatuses()
 	}
 	WriteHeaders(w, pb)
@@ -122,6 +139,13 @@ func GetTransitions(w http.ResponseWriter, req *http.Request) {
 // AbortTransitionID - abort transition by transitionID
 func AbortTransitionID(w http.ResponseWriter, req *http.Request) {
 	pb := GetUUIDFromVars("transitionID", req)
+
+	// Drain and close request body to ensure connection reuse
+	if (req.Body != nil) {
+		_, _ = io.Copy(io.Discard, req.Body)
+		req.Body.Close()
+	}
+
 	if pb.IsError {
 		WriteHeaders(w, pb)
 		return
