@@ -1,17 +1,17 @@
 // MIT License
-// 
+//
 // (C) Copyright [2022-2024] Hewlett Packard Enterprise Development LP
-// 
+//
 // Permission is hereby granted, free of charge, to any person obtaining a
 // copy of this software and associated documentation files (the "Software"),
 // to deal in the Software without restriction, including without limitation
 // the rights to use, copy, modify, merge, publish, distribute, sublicense,
 // and/or sell copies of the Software, and to permit persons to whom the
 // Software is furnished to do so, subject to the following conditions:
-// 
+//
 // The above copyright notice and this permission notice shall be included
 // in all copies or substantial portions of the Software.
-// 
+//
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 // FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
@@ -23,24 +23,24 @@
 package domain
 
 import (
-    "bytes"
-    "fmt"
-    "io/ioutil"
-    "net/http"
-    "os"
+	"bytes"
+	"fmt"
+	"io"
+	"net/http"
+	"os"
 	"sync"
-    "testing"
+	"testing"
 	"time"
 
-    "github.com/Cray-HPE/hms-certs/pkg/hms_certs"
-    "github.com/Cray-HPE/hms-power-control/internal/credstore"
-    "github.com/Cray-HPE/hms-power-control/internal/hsm"
-    pcsmodel "github.com/Cray-HPE/hms-power-control/internal/model"
-    "github.com/Cray-HPE/hms-power-control/internal/storage"
-    "github.com/Cray-HPE/hms-xname/xnametypes"
-    "github.com/sirupsen/logrus"
-    "github.com/stretchr/testify/suite"
-	trsapi "github.com/Cray-HPE/hms-trs-app-api/v2/pkg/trs_http_api"
+	"github.com/Cray-HPE/hms-certs/pkg/hms_certs"
+	"github.com/Cray-HPE/hms-power-control/internal/credstore"
+	"github.com/Cray-HPE/hms-power-control/internal/hsm"
+	pcsmodel "github.com/Cray-HPE/hms-power-control/internal/model"
+	"github.com/Cray-HPE/hms-power-control/internal/storage"
+	trsapi "github.com/Cray-HPE/hms-trs-app-api/v3/pkg/trs_http_api"
+	"github.com/Cray-HPE/hms-xname/xnametypes"
+	"github.com/sirupsen/logrus"
+	"github.com/stretchr/testify/suite"
 )
 
 type PwrStat_TS struct {
@@ -107,7 +107,8 @@ func glbInit() {
 		DLOCK.Init(tlogger)
 
 		domGlb.NewGlobals(nil, &TLOC_rf, nil, nil, svcClient, &sync.RWMutex{},
-			&running, &DSP, &HSM, vaultEnabled, &CS, &DLOCK, 20000, 1440)
+			&running, &DSP, &HSM, vaultEnabled, &CS, &DLOCK, 20000, 1440,
+			"power-status_test-pod")
 
 tlogger.Errorf("DLOCK: '%v', Globals: '%v'",DLOCK,domGlb)
 		domGlbInit = true
@@ -146,11 +147,24 @@ func doHTTP(url string, method string, pld []byte, auth *bAuth) ([]byte,int,erro
 	req.Header.Set("Content-Type","application/json")
 
 	rsp,perr := svcClient.Do(req)
+
 	if (perr != nil) {
+		// Always close response bodies
+		if rsp != nil && rsp.Body != nil {
+			_, _ = io.Copy(io.Discard, rsp.Body)
+			rsp.Body.Close()
+		}
+
 		return rdata,0,fmt.Errorf("Error performing http %s: %v",method,perr)
 	}
 
-	rdata,err = ioutil.ReadAll(rsp.Body)
+	rdata,err = io.ReadAll(rsp.Body)
+
+	// Always close response bodies
+	if rsp != nil && rsp.Body != nil {
+		rsp.Body.Close()
+	}
+
 	if (err != nil) {
 		return rdata,0,fmt.Errorf("Error reading http rsp body: %v",err)
 	}
@@ -273,7 +287,7 @@ func (suite *PwrStat_TS) Test_PowerStatusMonitor() {
 	t = suite.T()
 
 	//First initialize the power monitor
-	err := PowerStatusMonitorInit(&domGlb, (600*time.Second), tlogger, (5*time.Second), 30, 3)
+	err := PowerStatusMonitorInit(&domGlb, (600*time.Second), tlogger, (5*time.Second), 30, 3, 100, 2)
 	suite.Assert().Equal(nil,err,"PowerStatusMonitorInit() error: %v",err)
 
 	//Wait a while for the internal componant map to get updated.
