@@ -39,7 +39,7 @@ import (
 	"github.com/Cray-HPE/hms-power-control/internal/logger"
 	"github.com/Cray-HPE/hms-power-control/internal/model"
 	"github.com/Cray-HPE/hms-power-control/internal/storage"
-	trsapi "github.com/Cray-HPE/hms-trs-app-api/v2/pkg/trs_http_api"
+	trsapi "github.com/Cray-HPE/hms-trs-app-api/v3/pkg/trs_http_api"
 )
 
 var RFServerUrl string
@@ -275,7 +275,8 @@ func doSetup() error {
 	HSM.Init(&hsmGlob)
 
 	domainGlobals.NewGlobals(&BaseTRSTask, &TLOC_rf, &TLOC_svc, nil, nil,
-		rfClientLock, &Running, &DSP, &HSM, VaultEnabled, nil, nil, 20000, 1440)
+		rfClientLock, &Running, &DSP, &HSM, VaultEnabled, nil, nil,
+		20000, 1440, "power-cap_test-pod")
 	Init(&domainGlobals)
 	return nil
 }
